Funzionalità inizio tirocinio

+++ src/model/DAO/TirocinanteDAO.java
@@ -146,7 +146,7 @@ public class TirocinanteDAO {
                 "INSERT INTO TIROCINANTE(MATRICOLA, DATANASCITA, LUOGONASCITA, "
                     + "CITTADINANZA, RESIDENZA, CODICEFISCALE, TELEFONO, EMAIL) "
                     + "VALUES (?, ?, ?, ?, ?, ?, ?, ?)");
-            psTirocinante.setInt(1, tirocinante.getMatricola());
+            psTirocinante.setLong(1, tirocinante.getMatricola());
             //DateFormat format = new SimpleDateFormat("yyyy-MM-dd", Locale.ITALIAN);
             Date date = tirocinante.getDataNascita();
             psTirocinante.setDate(2, (java.sql.Date) date);
